client/requirements.txt
pyserial
termcolor
pyyaml
argparse
